Move actions to channel-nav component
The actions are navigation specific and don't belong in the application route.
