+++ app/routes/application.js
@@ -1,11 +1,9 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { isEmpty } from '@ember/utils';
 
 export default Route.extend({
   logger: service(),
   coms: service(),
-  storage: service('remotestorage'),
 
   beforeModel() {
     this._super(...arguments);
@@ -38,32 +36,8 @@ export default Route.extend({
           this.controller.toggleProperty(menuProp);
           break;
       }
-    },
-
-    openNewChannel(space) {
-      let channelName = window.prompt('Join channel');
-
-      if (isEmpty(channelName)) {
-        return;
-      }
-
-      if (space.get('protocol') === 'IRC' && !channelName.match(/^#/)) {
-        channelName = `#${channelName}`;
-      }
-      let channel = this.coms.createChannel(space, channelName);
-      this.storage.saveSpace(space);
-      this.transitionTo('space.channel', space, channel);
-    },
-
-    partChannel(space, channel) {
-      this.get('coms').removeChannel(space, channel.get('name'));
-
-      // Switch to last channel if the channel parted was currently open
-      if (channel.visible) {
-        let lastChannel = space.get('sortedChannels.lastObject');
-        this.transitionTo('space.channel', space, lastChannel);
-      }
     }
+
   }
 
 });
