+++ app/templates/application.hbs
@@ -7,7 +7,7 @@
 
     <div class="main">
       <nav id="channels">
-        {{channel-nav spaces=spaces joinChannel=(route-action "openNewChannel") partChannel=(route-action "partChannel")}}
+        {{channel-nav spaces=spaces}}
       </nav>
     </div>
   </div>
